requirements.txt
torch>=2.0
torchvision>=0.15
numpy
PyYAML
tqdm
matplotlib
